#27 - moves last month component to home module
