+++ feature/home/src/main/java/br/com/rbrthmn/home/ui/components/lastmonthdifferencecard/LastMonthDifferenceCardContract.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.screens.home.components.lastmonthdifferencecard
+package br.com.rbrthmn.home.ui.components.lastmonthdifferencecard
 
 import br.com.rbrthmn.ui.BaseViewModel
 import java.time.LocalDate
